requirements/development.txt
-r base.txt
pygraphviz==1.2
django-extensions==1.5.2
django-debug-toolbar==1.3.0
flake8==2.4.0
coverage==4.0.1
